Sped up the post processing and added GA
Reduced what was being passed to the distance matrices in an effort to reduce the time it takes to run. Tried parfor, but either the current setup is not suited to parfor (structs) or the memory cost is too high. Either way it is slightly faster now for what we want to do, but more work could be done with parfor in the future.